Include files in binary
Uses `go-bindata` (1)to generate a `bindata.go` file which
includes the file data in the binary.

Additionally we use `go-bindata-assetfs` (1) to provide the
data as static files in gin as well as `multitemplate` (3)
to load it into the template system.

Note that this commit also drastically changes the make system.

1: https://github.com/tmthrgd/go-bindata
2: https://github.com/elazarl/go-bindata-assetfs
3: https://github.com/gin-contrib/multitemplate